requirements.txt
Keras==2.2.4
numpy==1.16.1
opencv_python
tensorflow==1.12.3
Pillow==6.0.0
Flask==0.10.1
waitress